positional evaluation function

+++ src/Minimax.java
@@ -1,9 +1,27 @@
+import java.util.ArrayList;
+
 /**
  * Created by NSPACE on 11/18/2016.
  */
 public class Minimax {
 
-    public long minimax(long[] bitboards) {
+//    public long[] makeMinimaxMove() {
+//        //get child boards
+//        ArrayList<long[]> childBitBoards = getChildBitboards(bbSelf, bbEnemy, turn);
+//
+//        //return same board and exit if no moves can be made
+//        if (childBitBoards.size() == 0) {
+//            if (turn % 2 == 0) {
+//                return new long[]{bbSelf, bbEnemy};
+//            } else {
+//                return new long[]{bbEnemy, bbSelf};
+//            }
+//        }
+//
+//        //run minimax maxPlayer = false on children keep largest score board
+//    }
+
+    public int minimax(long[] bitboards) {
 
 
 //        01 function minimax(node, depth, maximizingPlayer)
